Extracting aspect logic to a separate component
